EDU-1027 Extract hook for isMounted flag
