+++ src/plugins/matching-cards/matching-cards-display.js
@@ -2,10 +2,11 @@ import classNames from 'classnames';
 import { SIZE } from './constants.js';
 import { Button, Tooltip } from 'antd';
 import { useTranslation } from 'react-i18next';
+import { useIsMounted } from '../../ui/hooks.js';
 import { ReloadOutlined } from '@ant-design/icons';
+import React, { useEffect, useState } from 'react';
 import FlipCard from '../../components/flip-card.js';
 import MatchingCardsTile from './matching-cards-tile.js';
-import React, { useEffect, useRef, useState } from 'react';
 import CheckIcon from '../../components/icons/general/check-icon.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
 import { getRandomizedTilesFromPairs } from './matching-cards-utils.js';
@@ -14,17 +15,10 @@ function MatchingCardsDisplay({ content }) {
   const { size, tilePairs, width } = content;
   const { t } = useTranslation('matchingCards');
 
-  const isMounted = useRef(false);
+  const isMounted = useIsMounted();
   const [tiles, setTiles] = useState([]);
   const [matchedTilePairKeys, setMatchedTilePairKeys] = useState([]);
   const [currentlyFlippedTiles, setCurrentlyFlippedTiles] = useState([]);
-
-  useEffect(() => {
-    isMounted.current = true;
-    return () => {
-      isMounted.current = false;
-    };
-  }, []);
 
   useEffect(() => {
     setMatchedTilePairKeys([]);
